// 6to Semestre/Aplicaciones Moviles II/Pokedex/src/app/models/PokemonDetail.ts
// Generated by https://quicktype.io

export interface PokemonDetail {
    id:              number;
    name:            string;
    typeDetail:      TypeDetail[];
    weight:          number;
    height:          number;
    base_experience: number;
    species:         Species;
    abilities:       Ability[];
    stats:           Stat[];
}

export interface Ability {
    ability:   PropertyName;
    is_hidden: boolean;
}

export interface PropertyName {
    name: string;
}

export interface Species {
    id:            number;
    speciesName:    PropertyName[];
    base_happiness: number;
    descriptions:  Descriptions[];
    capture_rate:   number;
    growthRate:      PropertyName;
    propertyName:     PropertyName;
    gender_rate:    number;
    hatch_counter:  number;
    eggGroups:      EggGroup[];
    dexnumbers:     Dexnumber[];
    evolutionChain: EvolutionChain;
}

export interface Dexnumber {
    pokedex_number: number;
    pokedex:        Pokedex;
}

export interface Pokedex {
    version_group: VersionGroup[];
    region:        PropertyName | null;
}

export interface VersionGroup {
    group: PropertyName;
}

export interface EggGroup {
    eggGroup: PropertyName;
}

export interface EvolutionChain {
    id:               number;
    evolutionSpecies: EvolutionSpecy[];
}

export interface EvolutionSpecy {
    pokemonId: number;
    name:      string;
    evolveFromId: number;
    evolution: Evolution[];
}

export interface Evolution {
    move:                 string;
    time_of_day:          string;
    turn_upside_down:     boolean;
    min_affection:        number;
    min_beauty:           number;
    min_happiness:        number;
    min_level:            number;
    needs_overworld_rain: boolean;
    heldItem:             string;
    item:                 PropertyName;
    trigger:              PropertyName;
}

export interface Stat {
    effort:    number;
    base_stat: number;
    stat:      PropertyName;
}

export interface TypeDetail {
    type: PropertyName;
}

export interface Descriptions {
    description: string;
}